requirements.txt
fastapi
pydantic
python-dotenv
openai
scipy
uvicorn
aiohttp
aiofiles
python-docx
PyPDF2